add webhook watching for Pods
Set up pieces for first webhook. The webhook will be built as its own go
binary (named 'webhook'). Update the Dockerfile to copy and build the
new webhook/main.go. Update the Makefile so that local builds also build
the webhook binary.

Generate the ValidatingWebhookConfiguration file (into config/webhook/),
and set up a separate set of files for kustomize to be able to generate
manifests for. Extend the Makefile's 'generate-manifests' to build the
webhook manifests and place them into coollers/webhook-manifests.yaml.
This set of manifests presently consists of a Deployment for running the
webhok binary, a Service to point to the Deployment's Pod, and the
VallidatingWebhookConfiguration generated by kubebuilder.

Signed-off-by: Joel Diaz <joel@mondoo.com>