add registry scanning fields for MondooAuditConfig
Extend MondooAuditConfig with a 'Registries' section.

It can hold a list of registry names/urls/endpoints that can be used to build a config to pass to the mondoo-client for scanning.
Rather than just a list of strings, each "registry" entry in the list is a distinct struct to allow more easily extending each entry with additional configuration fields in the future (things like scanning only specific repos, adding references to authentication for individual registries, etc).

Signed-off-by: Joel Diaz <joel@mondoo.com>