record mondoo-operator image as env var
The mondoo-operator is responsible for launching "itself" when running the webhook Pod. The operator needs a way to know which image it has been built for, so store this an the env var MONDOO_OPERATOR_IMAGE.

Use this env var when getting ready to deploy the webhook to set the container image properly.

Signed-off-by: Joel Diaz <joel@mondoo.com>