migrate to using new MondooOperatorConfig for metrics
Rather than put the metrics config into the MondooAuditConfig (which is
really for configuring monitoring-specific settings), create a new
MondooOperatorConfig CRD which is cluster-scoped which can be used to
configure operator-wide behavor of the mondoo-operator.

In a cluster with multiple MondooAuditConfigs, it makes no sense to have
one resource with metrics.enabled = true and a different one with
metrics.enabled = false. So just allow a single MondooOperatorConfig to
hold the cluster-wide metrics configuration for the mondoo-operator.

Take the existing ServiceMonitor handling code and call it from the new
mondoooperatorconfig controller.

Extend the MondooOperatorConfig status to hold a list of conditions, and
use this to communicate status for when metrics is enabled, but we
couldn't find Prometheus installed on the cluster.

The conditions handing is written so that a Condition only appears
initially if the Condition.Status is set to True. This means that if you
enable metrics, and Prometheus is found, there will be no
Condition[].Type = PrometheusMissing with .Status = False. Only when
Prometheus is missing will the condition be populated, and of course if
Prometheus transitions from Missing to Found, then the Condition will be
updated to show .Type = PrometheusMissing .Status = False.

Signed-off-by: Joel Diaz <joel@mondoo.com>