move to http metrics
Using a ServiceMonitor with the bearerTokenFile parameter set causes the
ServiceMonitor to be rejected by the OpenShift user monitoring stack (
https://github.com/operator-framework/operator-sdk/issues/4764 ).

As there is nothing sensitive in the mondoo-operator metrics, just
expose them directly to allow metrics to work under the built-in
OpenShift user metrics monitoring stack.

Add the ability to set some labels on the ServiceMonitor to allow a
functional metrics collection with an out-of-the-box prometheus deployed
as configured in
https://github.com/prometheus-community/helm-charts/tree/main/charts/kube-prometheus-stack
.

Change the kustomize generation so that the kube-rbac-proxy sidecar
container is no longer defined. It really only exists to protect
metrics. Introduce new Service to expose the new metrics ports. Patch
the default Deployment to expose the metrics port. A side benefit of
this is that you don't need to specify the container name when
displaying logs for mondoo-operator as there is now only a single
container.

Signed-off-by: Joel Diaz <joel@mondoo.com>