⭐️ Expose Metrics for Mondoo Operator (#186)
* Expose metrics for prometheus
* Added Status

Signed-off-by: Harsha <harshaisgud@gmail.com>

* migrate to using new MondooOperatorConfig for metrics

Rather than put the metrics config into the MondooAuditConfig (which is
really for configuring monitoring-specific settings), create a new
MondooOperatorConfig CRD which is cluster-scoped which can be used to
configure operator-wide behavor of the mondoo-operator.

In a cluster with multiple MondooAuditConfigs, it makes no sense to have
one resource with metrics.enabled = true and a different one with
metrics.enabled = false. So just allow a single MondooOperatorConfig to
hold the cluster-wide metrics configuration for the mondoo-operator.

Take the existing ServiceMonitor handling code and call it from the new
mondoooperatorconfig controller.

Extend the MondooOperatorConfig status to hold a list of conditions, and
use this to communicate status for when metrics is enabled, but we
couldn't find Prometheus installed on the cluster.

The conditions handing is written so that a Condition only appears
initially if the Condition.Status is set to True. This means that if you
enable metrics, and Prometheus is found, there will be no
Condition[].Type = PrometheusMissing with .Status = False. Only when
Prometheus is missing will the condition be populated, and of course if
Prometheus transitions from Missing to Found, then the Condition will be
updated to show .Type = PrometheusMissing .Status = False.

Signed-off-by: Joel Diaz <joel@mondoo.com>

* move to http metrics

Using a ServiceMonitor with the bearerTokenFile parameter set causes the
ServiceMonitor to be rejected by the OpenShift user monitoring stack (
https://github.com/operator-framework/operator-sdk/issues/4764 ).

As there is nothing sensitive in the mondoo-operator metrics, just
expose them directly to allow metrics to work under the built-in
OpenShift user metrics monitoring stack.

Add the ability to set some labels on the ServiceMonitor to allow a
functional metrics collection with an out-of-the-box prometheus deployed
as configured in
https://github.com/prometheus-community/helm-charts/tree/main/charts/kube-prometheus-stack
.

Change the kustomize generation so that the kube-rbac-proxy sidecar
container is no longer defined. It really only exists to protect
metrics. Introduce new Service to expose the new metrics ports. Patch
the default Deployment to expose the metrics port. A side benefit of
this is that you don't need to specify the container name when
displaying logs for mondoo-operator as there is now only a single
container.

Signed-off-by: Joel Diaz <joel@mondoo.com>

Co-authored-by: Joel Diaz <joel@mondoo.com>