fake some event data when running via act
It appears that act doesn't set up enough of a fake environment which
can cause some github action workflows that work on the real github to
fail when run through act.

Create some fake event data so that the checks against specific comments
in the spell-check workflow can succeed when run locally.

Signed-off-by: Joel Diaz <joel@mondoo.com>