pass Webhook mode down to Webhook Deployment
Take the MondooAuditConfig and use the .Spec.Webhooks.Mode value for setting the MONDOO_WEBHOOK_MODE environment variable in the Deployment for the webhook.

Use the value when registering the webhook so that the webhook handler can distinguish between modes.

Add test cases to validate that the mondoo-operator is in fact creating the webhook Deployment properly.

Restructure the webhook event handling for a future where we can call into the Scan Service.

Signed-off-by: Joel Diaz <joel@mondoo.com>