random cleanups
Explicitly set the mondoo-operator imagePullPolicy to IfNotPresent. The 'kustomize build | kubectl apply' process was resulting in an imagePullPolicy of 'Always' which makes local testing/development cumbersome.

Add the new 'skipContainerResolution' field to the sample MondooOperatorConfig so that someone can more easily toggle the operator's behavior without having to remember the name of that field.

In a similar vein, fill out the sample MondooAuditConfig with the fields showing the certificateConfig for the Webhooks section (with a reminder that cert-manager needs to be installed for the sample to work if webhooks are set to enable == true).

Signed-off-by: Joel Diaz <joel@mondoo.com>