scale back the resource requests for Pods
As CPU is a very compressible resource, we can ask for far less than 375m (which is really basically saying we need 1/3 of a full CPU). Bring that way back to 50m (5% of a CPU). Still allowing the Pod access to a full 1/2 a CPU in the event that a particular scan was CPU intensive.

For Memory, scale it back a bit from requesting 750M (which was 75% of the Pod Memory limit), down to 500M (50% of the limit). This should guarantee a minimum of .5GB with a bit of "overscheduling" the underlying cluster memory-wise).

Signed-off-by: Joel Diaz <joel@mondoo.com>