cleanup and wait for the Pods to come up (#254)
Just like in the regular mondoo-namespace test, wait for the Pods. Otherwise we risk racing past to the next part of e2e and failing because we didn't wait.

More importantly, we've seen Pods fail to come up due to lack of resources in the minikube environment. So clean up the MondooAuditConfig from the mondoo-operator namespace test to free up some resource for the not-mondoo-operator tests.

Signed-off-by: Joel Diaz <joel@mondoo.com>