⭐️ webhook mode of permissive/enforcing (#251)
* add field to control permissive/enforcing webhook behavior

Located in MondooAuditConfig.Spec.Webhooks.Mode.

It can be "", "permissive", or "enforcing". The empty string is handled internally as if "permissive were selected.

Take the MondooAuditConfig and use the .Spec.Webhooks.Mode value for setting the MONDOO_WEBHOOK_MODE environment variable in the Deployment for the webhook.

Explicitly set the mondoo-operator imagePullPolicy to IfNotPresent. The 'kustomize build | kubectl apply' process was resulting in an imagePullPolicy of 'Always' which makes local testing/development cumbersome.

* make sure to clear conditions when scanning is disabled

No point in reporting Degraded=True if the scanning isn't even enabled.

Signed-off-by: Joel Diaz <joel@mondoo.com>