deploy Scan API will token
As part of standing up the Scan API Deployment, create a Secret to hold the token used to authenticate between the webhook and the Scan API.

Limit our permissions related to Secrets to just Create/Delete. This means we just try to create the Secret and if it already exists, it means the token was generated/saved w/o us having to be able to read Secrets across Namespaces.

Use the new --token-file-path parameter to point to the data in the mounted Secret.

Signed-off-by: Joel Diaz <joel@mondoo.com>