have webhook call the Scan API
Now that the Scan API is ready to receive requests, link up the webhook to the Scan API via pkg/scanner. Re-use the same Secret with the token for the Scan API for the Webhook to also have the token when making the API calls.

Pass the location of the token, and the URL/Service to call for the Scan API.

Now that we are passing data to the webhook via CLI parameters, also pass the enforcement mode as a parameter and retire doing it via an env var.

Rework the testing for the scanner so that pieces of it can be used in the webhook testing.

Signed-off-by: Joel Diaz <joel@mondoo.com>