🌟 Scan on webhook handling (#303)
* factor out namespace fetching function
* deploy Scan API with token

As part of standing up the Scan API Deployment, create a Secret to hold the token used to authenticate between the webhook and the Scan API.

Limit our permissions related to Secrets to just Create/Delete. This means we just try to create the Secret and if it already exists, it means the token was generated/saved w/o us having to be able to read Secrets across Namespaces.

Use the new --token-file-path parameter to point to the data in the mounted Secret.

* have webhook call the Scan API

Now that the Scan API is ready to receive requests, link up the webhook to the Scan API via pkg/scanner. Re-use the same Secret with the token for the Scan API for the Webhook to also have the token when making the API calls.

Pass the location of the token, and the URL/Service to call for the Scan API.

Now that we are passing data to the webhook via CLI parameters, also pass the enforcement mode as a parameter and retire doing it via an env var.

Rework the testing for the scanner so that pieces of it can be used in the webhook testing.

* use kube-system Namespace UID as clusterID

* spell check

spell check is acting up. make it happy

* pass on resource-version label to scan

Signed-off-by: Joel Diaz <joel@mondoo.com>