⭐ token to service account (#323)
* token to service account

Allow specifying a MondooAuditConfig.Spec.MondooTokenSecretRef. If the Secret referenced in .Spec.MondooCredsSecretRef doesn't exists, and MondooTokenSecretRef does exist, use the token data to exchange for a Mondoo service account and save it into the MondooCredsSecretRef Secret.

Create a new mondooclient package (modeled after the scanner package) to wrap the interactions with the Mondoo API. Generate a mock client to allow testing of the various token/creds scenarios.

Add RBAC to the mondoo-operator ClusterRole to allow reading the Secrets named 'mondoo-client' and 'mondoo-token'. A user wishing to specify different Secret names will need to update the RBAC rules to allow mondoo-operator to read those alternative Secret names.

* unify scanner and mondooclient

Signed-off-by: Joel Diaz <joel@mondoo.com>