attempt an immediate CheckIn() after Register()
When trading a token for a Mondoo service account, don't wait for the periodic integration controller to run and perform the first CheckIn(). Just make an attempt right after saving the Secret with the Mondoo service account.

Update test cases to expect the additional CheckIn() API call, and restructure the Integration controller to more easily allow perfoming a CheckIn() using only the Integration MRN and the Mondoo service account data.

Signed-off-by: Joel Diaz <joel@mondoo.com>