no more early exits during Reconcile()
Remove the last remaining places where the Reconcile() loop would exit early in non-error cases.

Additionally, rework how the workload codepath determines when an update to the Pods is needed (for example if the contents of the ConfigMap change). Instead of indicating that a change was made to the ConfigMap and then putting a timestamp on the Deployment to force a re-deploy, just compute the hash of the ConfigMap contents and include that in the Deployment as an annoation. Now if the ConfigMap contents change, the hash will change, and the workload scanning Pod will be re-deployed even if an error is encountered between updating the ConfigMap and updating the Deployment.

Signed-off-by: Joel Diaz <joel@mondoo.com>