create separate default Node resource requirements
Having a 400m CPU request for the Node scanning Jobs is excessive for an hourly node scan where the completion time is not time sensitive.

Rather than have the Node resource defaults be shared with the singleton ScanAPI Deployment, allow for the Node default resource requirements to be different. Lower the CPU requests for the Node Jobs down to 50m so that scheduling a node scan doesn't request such a high amount of CPU on each node.

Signed-off-by: Joel Diaz <joel@mondoo.com>