stop labeling resources with key 'control-plane'
First of all, using 'control-plane' as a label is misleading as the Mondoo Operator is not part of any k8s control plane.

Second, a ValidatingWebhookConfiguration installed on an AKS cluster (which is used to be notified about changes to the cluster so that those resources are scanned), will be modified to ignore resources with the label 'control-plane'. This means Mondoo Operator on AKS would behave differently than on other k8s environments.

Stop using the 'control-plane' label and replace it with app=mondoo-operator and app=mondoo where appropriate.

This change means that upgrading cannot be done with a simple `kubectl apply` as changing the selection labels on the operator Deployment isn't allowed. The upgrade process is changed from a `kubectl apply` of all the manifests to first deleting the operator Deployment, then running the normal apply.

Signed-off-by: Joel Diaz <joel@mondoo.com>