🧹 No more 'control-plane' labeling (#460)
* stop labeling resources with key 'control-plane'

First of all, using 'control-plane' as a label is misleading as the Mondoo Operator is not part of any k8s control plane.

Second, a ValidatingWebhookConfiguration installed on an AKS cluster (which is used to be notified about changes to the cluster so that those resources are scanned), will be modified to ignore resources with the label 'control-plane'. This means Mondoo Operator on AKS would behave differently than on other k8s environments.

Stop using the 'control-plane' label and replace it with app=mondoo-operator and app=mondoo where appropriate.

This change means that upgrading cannot be done with a simple `kubectl apply` as changing the selection labels on the operator Deployment isn't allowed. The upgrade process is changed from a `kubectl apply` of all the manifests to first deleting the operator Deployment, then running the normal apply.

Signed-off-by: Joel Diaz <joel@mondoo.com>

* add AKS-specific label

this allows the webhook to watch Namespaces with the label 'control-plane'.

Also fix up the updating of the webhook Deployment as the changes in labels means we need to delete/create instead of update.

Signed-off-by: Joel Diaz <joel@mondoo.com>

* update test cases

now that we delete/create a Deployment, need to ensure the test cases create objects that won't get updated

Signed-off-by: Joel Diaz <joel@mondoo.com>

* remove some early dev files

Signed-off-by: Joel Diaz <joel@mondoo.com>

* address PR feedback

Signed-off-by: Joel Diaz <joel@mondoo.com>

* remove more cruft unrelated to this PR

Signed-off-by: Joel Diaz <joel@mondoo.com>

* watch for new-style label

Signed-off-by: Joel Diaz <joel@mondoo.com>

* delete unused file
add upgrade instructions for kubectl

Signed-off-by: Ivan Milchev <ivan@mondoo.com>

* added extra labels that do not overlap with Helm internals

Signed-off-by: Ivan Milchev <ivan@mondoo.com>

* added OLM installation upgrade manual

Signed-off-by: Ivan Milchev <ivan@mondoo.com>

* removed extra label for deployments

Signed-off-by: Ivan Milchev <ivan@mondoo.com>

Co-authored-by: Ivan Milchev <ivan@mondoo.com>