detect OpenShift differently
The previous approach requires the mondoo-operator to have ClusterRole
permissions to Get() a CustomResourceDefinition. While we can easily
just add that permission to our pre-defined ClusterRole RBAC, there is
an option that doesn't require the additional permissions.

Extract the Prometheus detection code into a shared tool and use the
APIService detection to determine OpenShift vs non-OpenShift.

Also, we were printing an error when the mondoo-operator was running
into an error, but just kept running with 'isOpenShift' == false causing
issues at runtime.

Signed-off-by: Joel Diaz <joel@mondoo.com>