check for required APIs at startup
Rather than fail later while the operator is up and running, just check
whether some of the resources we depend on to operator properly exist
in the cluster.

For the first pass, just check that Jobs exist (to prove that our
checks actually work) and whether CronJobs exist (to catch the case of
an EKS cluster running k8s 1.20 which lacks CronJobs v1 (but happens to
have CronJobs v1beta1)).

Signed-off-by: Joel Diaz <joel@mondoo.com>