attempt garbage collection after successful scans
Introduce a 'garbage-collect' subcommand to ease testing/development,
and to allow the 'k8s-scan' subcommand to use this for garbage
collection purposes.

When a scan has run cleanly, make a GarbageCollectAssets() call for
resources that are older than 2 hours (for k8s assets), and older
than 2 days (for container-image assets).

For now, hide the functionality behind the
FEATURE_ENABLE_GARBAGE_COLLECTION feature flag.

Signed-off-by: Joel Diaz <joel@mondoo.com>