🧹 Re-work PR workflows
This will bring back the integration and security tests into the PR.

It will also allow us dependabot and fork PRs. But we need to apply a label,
before the tests start.

The label is also directly removed after the workflow starts, to prevent later
commits from just starting the tests again.

Signed-off-by: Christian Zunker <christian@mondoo.com>